Made ReplicaManagerPacket and PacketParameters serializable Moved PacketParametersHandler to replica_manager
