+++ CORBA_DFRS/src/replica_manager/ReplicaAliveHandler.java
@@ -7,7 +7,6 @@ import java.net.InetAddress;
 import java.net.SocketException;
 
 import replica_manager_packet.PacketParameters;
-import replica_manager_packet.PacketParametersHandler;
 import replica_manager_packet.ReplicaAliveOperation;
 import replica_manager_packet.ReplicaAliveReply;
 import udp.UdpHelper;
