+++ CORBA_DFRS/src/replica_manager_packet/ReplicaRebootOperation.java
@@ -1,5 +1,6 @@
 package replica_manager_packet;
 
-public class ReplicaRebootOperation {
+public class ReplicaRebootOperation extends PacketParameters {
+	private static final long serialVersionUID = 1L;
 	// TODO : Add required parameters
 }
